+++ dev/lwpkt_opts.h
@@ -37,6 +37,7 @@
 #define LWPKT_CFG_USE_ADDR      2
 #define LWPKT_CFG_ADDR_EXTENDED 2
 #define LWPKT_CFG_USE_CMD       2
+#define LWPKT_CFG_CMD_EXTENDED  2
 #define LWPKT_CFG_USE_CRC       2
 #define LWPKT_CFG_CRC32         2
 #define LWPKT_CFG_USE_FLAGS     2