+++ lwpkt/src/include/lwpkt/lwpkt.h
@@ -159,7 +159,7 @@ typedef struct lwpkt {
         uint32_t flags; /*!< Custom flags */
 #endif                  /* LWPKT_CFG_USE_FLAGS || __DOXYGEN__ */
 #if LWPKT_CFG_USE_CMD || __DOXYGEN__
-        uint8_t cmd;  /*!< Command packet */
+        uint32_t cmd; /*!< Command packet */
 #endif                /* LWPKT_CFG_USE_CMD || __DOXYGEN__ */
         size_t len;   /*!< Number of bytes to receive */
         size_t index; /*!< General index variable for multi-byte parts of packet */
@@ -177,7 +177,7 @@ lwpktr_t lwpkt_write(lwpkt_t* pkt,
                      uint32_t flags,
 #endif /* LWPKT_CFG_USE_FLAGS || __DOXYGEN__ */
 #if LWPKT_CFG_USE_CMD || __DOXYGEN__
-                     uint8_t cmd,
+                     uint32_t cmd,
 #endif /* LWPKT_CFG_USE_CMD || __DOXYGEN__ */
                      const void* data, size_t len);
 lwpktr_t lwpkt_reset(lwpkt_t* pkt);
@@ -185,12 +185,13 @@ lwpktr_t lwpkt_process(lwpkt_t* pkt, uint32_t time);
 lwpktr_t lwpkt_set_evt_fn(lwpkt_t* pkt, lwpkt_evt_fn evt_fn);
 
 /* Functions available as conditional build */
-void lwpkt_set_crc_enabled(lwpkt_t* pkt, uint8_t enable);
-void lwpkt_set_crc32_enabled(lwpkt_t* pkt, uint8_t enable);
 void lwpkt_set_addr_enabled(lwpkt_t* pkt, uint8_t enable);
 void lwpkt_set_addr_extended_enabled(lwpkt_t* pkt, uint8_t enable);
-void lwpkt_set_cmd_enabled(lwpkt_t* pkt, uint8_t enable);
 void lwpkt_set_flags_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_cmd_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_cmd_extended_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_crc_enabled(lwpkt_t* pkt, uint8_t enable);
+void lwpkt_set_crc32_enabled(lwpkt_t* pkt, uint8_t enable);
 
 /**
  * \brief           Get address from where packet was sent
@@ -225,7 +226,7 @@ void lwpkt_set_flags_enabled(lwpkt_t* pkt, uint8_t enable);
  * \param[in]       pkt: LwPKT instance
  * \return          Command data field
  */
-#define lwpkt_get_cmd(pkt)       (uint8_t)(((pkt) != NULL) ? ((pkt)->m.cmd) : 0)
+#define lwpkt_get_cmd(pkt)       (uint32_t)(((pkt) != NULL) ? ((pkt)->m.cmd) : 0)
 
 /**
  * \brief           Get packet flags